+++ src/main/java/controller/ResourceController.java
@@ -1,24 +1,16 @@
 package controller;
 
 import annotation.GetMapping;
-import db.Database;
-import model.Post;
 import model.User;
 import request.HttpRequest;
 import response.HttpResponse;
 import response.HttpResponseBuilder;
 import response.HttpResponseStatus;
 import session.SessionManager;
+import view.ViewResolver;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.time.format.DateTimeFormatter;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 public class ResourceController {
     Map<String, String> responseHeaders = new HashMap<>();
@@ -38,48 +30,20 @@ public class ResourceController {
     }
     @GetMapping(url = "/resources")
     public HttpResponse getResources(HttpRequest request) {
-        String url;
-        if (request.getUrl().endsWith(".html")) {
-            url = "src/main/resources/templates" + request.getUrl();
-        } else {
-            url = "src/main/resources/static" + request.getUrl();
+        Map<String, Object> model = new HashMap<>();
+        if(request.isAuth()) {
+            User loginUser = sessionManager.getUserBySessionId(request);
+            model.put("name", loginUser.getName());
         }
-        File file = new File(url);
-        if (file.isFile()) {
-            try (FileInputStream fis = new FileInputStream(file)) {
-                byte[] body = fis.readAllBytes();
 
-                User loginUser = sessionManager.getUserBySessionId(request);
-                if (loginUser != null && url.endsWith(".html")) {
-                    String content = new String(body, "UTF-8");
-                    String replacedBody = replaceTag(content, "로그인", loginUser.getName());
-                    body = replacedBody.getBytes("UTF-8");
-                }
-
-                if (request.getUrl().equals("/index.html")) {
-                    String content = new String(body, "UTF-8");
-                    String replacedBody = replaceTag(content, "<ul class=\"list\"></ul>", getContentListHtml(Database.getPostList()));
-                    body = replacedBody.getBytes("UTF-8");
-                }
-
-
-                responseHeaders.put(CONTENT_TYPE, getContentType(url));
-                responseHeaders.put(CONTENT_LENGTH, String.valueOf(body.length));
-                return new HttpResponseBuilder().status(HttpResponseStatus.OK)
-                        .headers(responseHeaders)
-                        .body(body)
-                        .build();
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-        } else {
-            responseHeaders.put(CONTENT_TYPE, "text/html;charset=utf-8");
-            return new HttpResponseBuilder().status(HttpResponseStatus.NOT_FOUND)
-                    .headers(responseHeaders)
-                    .build();
-        }
+        byte[] body = ViewResolver.resolve(request.getUrl(), request.isAuth(), model);
+        responseHeaders.put(CONTENT_TYPE, getContentType(request.getUrl()));
+        responseHeaders.put(CONTENT_LENGTH, String.valueOf(body.length));
+        return new HttpResponseBuilder().status(HttpResponseStatus.OK)
+                .headers(responseHeaders)
+                .body(body)
+                .build();
     }
-
     private String getContentType(String url) {
         String extension = url.substring(url.lastIndexOf(".") + 1);
         String result = contentType.get(extension);
@@ -87,53 +51,5 @@ public class ResourceController {
             return "font/" + extension;
         }
         return result;
-    }
-
-    private static String replaceTag(String content, String targetWord, String replacement) {
-        // 정규 표현식 패턴 생성
-        String regex = "\\Q" + targetWord + "\\E";  // 정규 표현식 특수 문자를 이스케이프
-        Pattern pattern = Pattern.compile(regex, Pattern.DOTALL);
-
-        // 패턴과 일치하는 단어 찾기
-        Matcher matcher = pattern.matcher(content);
-
-        // 대체된 내용으로 문자열 빌드
-        StringBuffer result = new StringBuffer();
-        while (matcher.find()) {
-            matcher.appendReplacement(result, Matcher.quoteReplacement(replacement));
-        }
-        matcher.appendTail(result);
-
-        return result.toString();
-    }
-
-    private String getContentListHtml(Map<Long, Post> postList) {
-        StringBuilder sb = new StringBuilder();
-        sb.append("<ul class=\"list\">");
-        for (Post post : postList.values()) {
-            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
-            String createTime = post.getCreateTime().format(formatter);
-
-            sb.append("<li>");
-            sb.append("<div class=\"wrap\">");
-            sb.append("<div class=\"main\">");
-            sb.append("<strong class=\"subject\">");
-            sb.append("<a href=\"./post/detail?id=" + post.getId() + "\">" + post.getTitle() + "</a>");
-            sb.append("</strong>");
-            sb.append("<div class=\"auth-info\">");
-            sb.append("<i class=\"icon-add-comment\"></i>");
-            sb.append("<span class=\"time\">"+createTime+"</span>");
-            sb.append("<a href=\"./user/profile.html\" class=\"author\">"+ post.getWriter() +"</a>");
-            sb.append("</div>");
-            sb.append("<div class=\"reply\" title=\"댓글\">");
-            sb.append("<i class=\"icon-reply\"></i>");
-            sb.append("<span class=\"point\">8</span>");
-            sb.append("</div>");
-            sb.append("</div>");
-            sb.append("</div>");
-            sb.append("</li>");
-        }
-        sb.append("</ul>");
-        return sb.toString();
     }
 }
