+++ src/main/java/controller/UserController.java
@@ -3,7 +3,6 @@ package controller;
 import annotation.GetMapping;
 import annotation.PostMapping;
 import annotation.RequestParam;
-import db.Database;
 import model.User;
 import request.HttpRequest;
 import response.HttpResponse;
@@ -12,15 +11,10 @@ import response.HttpResponseStatus;
 import service.UserJoinService;
 import service.UserLoginService;
 import session.SessionManager;
+import view.ViewResolver;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 public class UserController {
     Map<String, String> responseHeaders = new HashMap<>();
@@ -69,72 +63,25 @@ public class UserController {
 
     @GetMapping(url = "/user/list.html")
     public HttpResponse getList(HttpRequest request) {
-        User loginUser = sessionManager.getUserBySessionId(request);
-        if (loginUser == null) {
-            responseHeaders.put(LOCATION, "/user/login.html");
-            return new HttpResponseBuilder().status(HttpResponseStatus.FOUND)
-                    .headers(responseHeaders).build();
+        Map<String, Object> model = new HashMap<>();
+
+        if(!request.isAuth()) {
+            HttpResponse response = new HttpResponse();
+            responseHeaders.put("location", "/user/login.html");
+            response.setResponse(HttpResponseStatus.FOUND, null, responseHeaders);
+            return response;
         } else {
-            Collection<User> users = Database.findAll();
+            User loginUser = sessionManager.getUserBySessionId(request);
+            model.put("name", loginUser.getName());
 
-            File file = new File("src/main/resources/templates/user/list.html");
-            try(FileInputStream fis = new FileInputStream(file)) {
-                byte[] body = fis.readAllBytes();
-
-                String content = new String(body, "UTF-8");
-                String replacedBody = replaceWord(content, "로그인", loginUser.getName());
-                replacedBody = replaceWord(replacedBody, "<tbody></tbody>", makeListHtml(users));
-                body = replacedBody.getBytes("UTF-8");
-
-                responseHeaders.put(CONTENT_TYPE, "text/html; charset=utf-8");
-                responseHeaders.put(CONTENT_LENGTH, String.valueOf(body.length));
-                return new HttpResponseBuilder().status(HttpResponseStatus.OK)
-                        .headers(responseHeaders)
-                        .body(body)
-                        .build();
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
+            byte[] body = ViewResolver.resolve(request.getUrl(), request.isAuth(), model);
+            responseHeaders.put(CONTENT_TYPE, "text/html; charset=utf-8");
+            responseHeaders.put(CONTENT_LENGTH, String.valueOf(body.length));
+            return new HttpResponseBuilder().status(HttpResponseStatus.OK)
+                    .headers(responseHeaders)
+                    .body(body)
+                    .build();
         }
 
-    }
-
-    private String makeListHtml(Collection<User> users) {
-        StringBuilder sb = new StringBuilder();
-        int sequence = 0;
-        sb.append("<tbody>");
-        for (User user : users) {
-            sb.append("<tr>");
-            sb.append("<th scope=\"row\">");
-            sb.append(++sequence);
-            sb.append("</th> <td>");
-            sb.append(user.getUserId());
-            sb.append("</th> <td>");
-            sb.append(user.getName());
-            sb.append("</th> <td>");
-            sb.append(user.getEmail());
-            sb.append("</td><td><a href=\"#\" class=\"btn btn-success\" role=\"button\">수정</a></td>");
-            sb.append("</tr>");
-        }
-        sb.append("</tbody>");
-        return sb.toString();
-    }
-
-    private static String replaceWord(String content, String targetWord, String replacement) {
-        // 정규 표현식 패턴 생성
-        String regex = "\\Q" + targetWord + "\\E";  // 정규 표현식 특수 문자를 이스케이프
-        Pattern pattern = Pattern.compile(regex, Pattern.DOTALL);
-
-        // 패턴과 일치하는 단어 찾기
-        Matcher matcher = pattern.matcher(content);
-
-        // 대체된 내용으로 문자열 빌드
-        StringBuffer result = new StringBuffer();
-        while (matcher.find()) {
-            matcher.appendReplacement(result, Matcher.quoteReplacement(replacement));
-        }
-        matcher.appendTail(result);
-
-        return result.toString();
     }
 }
