+++ src/main/java/utils/HttpRequestUtils.java
@@ -1,6 +1,7 @@
 package utils;
 
 import request.HttpRequest;
+import session.SessionManager;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -12,6 +13,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class HttpRequestUtils {
+    private static final SessionManager sessionManager = new SessionManager();
 
     public static HttpRequest makeHttpRequest(InputStream in) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(in, StandardCharsets.UTF_8));
@@ -19,6 +21,11 @@ public class HttpRequestUtils {
         parseRequestLine(br, request);
         parseHeader(br, request);
         parseBody(request, br);
+        if(isLogin(request)) {
+            request.setAuth(true);
+        } else {
+            request.setAuth(false);
+        }
 
         // POST Form 방식 처리 구현
         if (request.getMethod().equals("POST") && request.getHeaders().get("Content-Type").equals("application/x-www-form-urlencoded") && !request.getUrl().equals("/user/logout")) {
@@ -104,4 +111,7 @@ public class HttpRequestUtils {
         return params;
     }
 
+    private static boolean isLogin(HttpRequest request) {
+        return sessionManager.getUserBySessionId(request) != null;
+    }
 }
