+++ src/main/resources/templates/index.html
@@ -96,7 +96,6 @@
                 </ul>
               </div>
               <div class="col-md-3 qna-write">
-                  <a href="./qna/form.html" class="btn btn-primary pull-right" role="button">질문하기</a>
                   <a href="./post" class="btn btn-primary pull-right" role="button" style="margin-right: 10px">글쓰기</a>
               </div>
           </div>
